Working win32 window and changed gameinterface to int

+++ source/Game.cpp
@@ -16,7 +16,7 @@ void Game::Initialize(RenderComponentInterface* p_renderComponentInterface)
 
 }
 
-void Game::Update(double p_deltaTime, float p_mousePositionX, float p_mousePositionY /* add keyboard parameters here*/)
+void Game::Update(double p_deltaTime, int p_mousePositionX, int p_mousePositionY /* add keyboard parameters here*/)
 {
 	m_renderComponentInterface->Render();
 }
